chore: move middleware tests closer to sources
